email: Move "`Into<BoxedAppError`" impl to `util/errors`
